+++ crawl-ref/source/env.h
@@ -60,7 +60,7 @@ public:
     // 5 turns).  Set to 0 to stop random generation.
     int spawn_random_rate;
 
-    double elapsed_time; // used during level load
+    long elapsed_time; // used during level load
 
     // Number of turns the player has spent on this level.
     int turns_on_level;
